the "find us here" section has clickable links
I thought it made sense for the github and email sections to be clickable links so that people can actually "find us" there. Thus I have edited the index.html file to wrap those things in <a> elements and I have edited the styles.css file to make the links consistent with the site's other links.